fix(overview-map): fixed loading the overview map
Closes #302
