+++ packages/geoview-loader/package.json
@@ -38,20 +38,8 @@
         ]
     },
     "dependencies": {
-        "@babel/core": "^7.17.0",
-        "@babel/runtime": "^7.17.0",
-        "@babel/runtime-corejs3": "^7.17.0",
-        "prop-types": "^15.8.1",
-        "domhandler": "^4.3.0",
-        "eventemitter3": "^4.0.7",
-        "html-react-parser": "^1.4.6",
-        "immutable": "^4.0.0",
-        "lodash": "^4.17.21",
-        "query-string": "^7.1.0",
         "react": "^17.0.2",
         "react-dom": "^17.0.2",
-        "screenfull": "^6.0.1",
-        "leaflet": "^1.7.1",
         "geoview-core": "workspace:~0.1.0",
         "geoview-details-panel": "workspace:~1.0.0",
         "geoview-overview-map": "workspace:*",
@@ -60,6 +48,9 @@
     },
     "devDependencies": {
         "@babel/cli": "^7.17.0",
+        "@babel/core": "^7.17.0",
+        "@babel/runtime": "^7.17.0",
+        "@babel/runtime-corejs3": "^7.17.0",
         "@babel/plugin-proposal-class-properties": "^7.16.7",
         "@babel/plugin-proposal-decorators": "^7.17.0",
         "@babel/plugin-proposal-do-expressions": "^7.16.7",
@@ -92,6 +83,8 @@
         "@typescript-eslint/eslint-plugin": "^5.10.2",
         "@typescript-eslint/parser": "^5.10.2",
         "@babel/eslint-parser": "^7.17.0",
+        "lodash": "^4.17.21",
+        "prop-types": "^15.8.1",
         "babel-jest": "^27.4.6",
         "babel-loader": "^8.2.3",
         "babel-plugin-import": "^1.13.3",
