+++ packages/geoview-loader/public/templates/usecases.html
@@ -88,7 +88,7 @@
                 { 'name': 'GeoJSON - point', 'url': './geojson/points.json', 'type': 'geoJSON' },
                 { 'name': 'GeoJSON - polygon', 'url': './geojson/polygons.json', 'type': 'geoJSON' },
                 { 'name': 'Geomet', 'url': 'https://geo.weather.gc.ca/geomet', 'type': 'ogcWMS', 'entries': 'RAQDPS-FW.CE_PM2.5-DIFF-YAvg' }
-            ]}"
+            ], 'plugins': ['overviewMap']}"
             ></div>
             <p>Use to display a map like a thumbnail with minimun functions.</p>
             <p><strong>Configuraion Snippet:</strong></p>
@@ -100,7 +100,7 @@
             <div
                 id="UC3"
                 class="llwp-map"
-                data-leaflet="{ 'name': 'LCC_SHADED', 'projection': 3978, 'zoom': 4, 'center': [60,-100], 'language': 'en-CA','basemapOptions': { 'id': 'simple', 'shaded': true, 'labeled': false }, 'layers': [] }"
+                data-leaflet="{ 'name': 'LCC_SHADED', 'projection': 3978, 'zoom': 4, 'center': [60,-100], 'language': 'en-CA','basemapOptions': { 'id': 'simple', 'shaded': true, 'labeled': false }, 'layers': [], 'plugins': ['overviewMap'] }"
             ></div>
             <p>Description</p>
             <p><strong>Configuraion Snippet:</strong></p>
@@ -112,7 +112,7 @@
             <div
                 id="UC4"
                 class="llwp-map"
-                data-leaflet="{ 'name': 'LCC_SHADED', 'projection': 3978, 'zoom': 4, 'center': [60,-100], 'language': 'en-CA', 'basemapOptions': { 'id': 'simple', 'shaded': true, 'labeled': false }, 'layers': [] }"
+                data-leaflet="{ 'name': 'LCC_SHADED', 'projection': 3978, 'zoom': 4, 'center': [60,-100], 'language': 'en-CA', 'basemapOptions': { 'id': 'simple', 'shaded': true, 'labeled': false }, 'layers': [], 'plugins': ['overviewMap'] }"
             ></div>
             <p>Description</p>
             <p><strong>Configuraion Snippet:</strong></p>
