+++ packages/geoview-overview-map/src/overview-map.tsx
@@ -1,280 +1,21 @@
-// eslint-disable-next-line import/no-unresolved
-import { Basemap } from "geoview-core/src/geo/layer/basemap/basemap";
-import { LEAFLET_POSITION_CLASSES } from "geoview-core/src/geo/utils/constant";
+import { Cast } from "geoview-core";
 
-import { Cast } from "geoview-core/src/core/types/cgpv-types";
+import { MinimapBounds } from "./minimap-bounds";
+import { MinimapToggle } from "./minimap-toggle";
 
-// get the window object
-const w = window as any;
-
-const MINIMAP_SIZE = {
+export const MINIMAP_SIZE = {
   width: "150px",
   height: "150px",
 };
 
-/**
- * Interface for overview map properties
- */
-interface OverviewProps {
-  id: string;
-  crs: Object;
-  language: string;
-  zoomFactor: number;
-}
-
-/**
- * Interface for bound polygon properties
- */
-interface MiniboundProps {
-  parentId: string;
-  parentMap: L.Map;
-  minimap: L.Map;
-  zoomFactor: number;
-}
-
-/**
- * Interface for the minimap toggle properties
- */
-interface MinimapToggleProps {
-  parentId: string;
-  minimap: L.Map;
-}
-
-/**
- * Create a toggle element to expand/collapse the overview map
- * @param {MinimapToggleProps} props toggle properties
- * @return {JSX.Element} the toggle control
- */
-function MinimapToggle(props: MinimapToggleProps): JSX.Element {
-  const { parentId, minimap } = props;
-
-  // access the cgpv object from the window object
-  const cgpv = w["cgpv"];
-
-  // access the api calls
-  const { api, react, leaflet, ui, useTranslation } = cgpv;
-
-  const { DomEvent } = leaflet;
-
-  // get event names
-  const EVENT_NAMES = api.eventNames;
-
-  const { useState, useEffect, useRef } = react;
-
-  const divRef = useRef(null);
-
-  const { t } = useTranslation();
-
-  const [status, setStatus] = useState(true);
-
-  // get available elements
-  const { IconButton, ChevronLeftIcon } = ui.elements;
-
-  const useStyles = ui.makeStyles((theme: any) => ({
-    toggleBtn: {
-      transform: "rotate(45deg)",
-      color: theme.palette.primary.contrastText,
-      zIndex: theme.zIndex.tooltip,
-    },
-    toggleBtnContainer: {
-      zIndex: theme.zIndex.tooltip,
-    },
-    minimapOpen: {
-      transform: "rotate(-45deg)",
-    },
-    minimapClosed: {
-      transform: "rotate(135deg)",
-    },
-  }));
-
-  const classes = useStyles();
-
-  const theme = ui.useTheme();
-
-  /**
-   * Toggle overview map to show or hide it
-   * @param e the event being triggered on click
-   */
-  function toggleMinimap(): void {
-    setStatus(!status);
-
-    if (status) {
-      const buttonSize = theme.overrides?.button?.size;
-      // decrease size of overview map to the size of the toggle btn
-      minimap.getContainer().style.width = buttonSize.width as string;
-      minimap.getContainer().style.height = buttonSize.height as string;
-    } else {
-      // restore the size of the overview map
-      minimap.getContainer().style.width = MINIMAP_SIZE.width;
-      minimap.getContainer().style.height = MINIMAP_SIZE.height;
-    }
-
-    // trigger a new event when overview map is toggled
-    api.event.emit(EVENT_NAMES.EVENT_OVERVIEW_MAP_TOGGLE, parentId, {
-      status,
-    });
-  }
-
-  useEffect(() => {
-    DomEvent.disableClickPropagation(Cast<HTMLElement>(divRef.current));
-  }, []);
-
-  return (
-    <div
-      ref={divRef}
-      className={`${LEAFLET_POSITION_CLASSES.topright} ${classes.toggleBtnContainer}`}
-    >
-      <IconButton
-        className={`leaflet-control ${classes.toggleBtn} ${
-          !status ? classes.minimapOpen : classes.minimapClosed
-        }`}
-        style={{
-          margin: `-${theme.spacing(3)}`,
-          padding: 0,
-          height: "initial",
-          minWidth: "initial",
-        }}
-        aria-label={t("mapctrl.overviewmap.toggle")}
-        onClick={toggleMinimap}
-        size="large"
-      >
-        <ChevronLeftIcon />
-      </IconButton>
-    </div>
-  );
-}
-
-/**
- * Create and update the bound polygon of the parent's map extent
- * @param {MiniboundProps} props bound properties
- */
-function MinimapBounds(props: MiniboundProps) {
-  const { parentId, parentMap, zoomFactor, minimap } = props;
-
-  // access the cgpv object from the window object
-  const cgpv = w["cgpv"];
-
-  // access the api calls
-  const { api, react, reactLeaflet, reactLeafletCore, ui, useTranslation } =
-    cgpv;
-
-  const { useMapEvent } = reactLeaflet;
-
-  const { useEventHandlers } = reactLeafletCore;
-
-  // get event names
-  const EVENT_NAMES = api.eventNames;
-
-  const { useState, useEffect, useCallback, useMemo } = react;
-
-  const [toggle, setToggle] = useState(false);
-
-  // Clicking a point on the minimap sets the parent's map center
-  const onClick = useCallback(
-    (e) => {
-      parentMap.setView(e.latlng, parentMap.getZoom());
-    },
-    [parentMap]
-  );
-  useMapEvent("click", onClick);
-
-  // Keep track of bounds in state to trigger renders
-  const [bounds, setBounds] = useState({
-    height: 0,
-    width: 0,
-    top: 0,
-    left: 0,
-  });
-
-  function updateMap(): void {
-    // Update the minimap's view to match the parent map's center and zoom
-    const newZoom =
-      parentMap.getZoom() - zoomFactor > 0
-        ? parentMap.getZoom() - zoomFactor
-        : 0;
-
-    minimap.flyTo(parentMap.getCenter(), newZoom);
-
-    // Set in timeout the calculation to create the bound so parentMap getBounds has the updated bounds
-    setTimeout(() => {
-      minimap.invalidateSize();
-      const pMin = minimap.latLngToContainerPoint(
-        parentMap.getBounds().getSouthWest()
-      );
-      const pMax = minimap.latLngToContainerPoint(
-        parentMap.getBounds().getNorthEast()
-      );
-      setBounds({
-        height: pMin.y - pMax.y,
-        width: pMax.x - pMin.x,
-        top: pMax.y,
-        left: pMin.x,
-      });
-    }, 500);
-  }
-
-  useEffect(() => {
-    updateMap();
-
-    // listen to API event when the overview map is toggled
-    api.event.on(
-      EVENT_NAMES.EVENT_OVERVIEW_MAP_TOGGLE,
-      (payload: any) => {
-        if (payload && parentId === payload.handlerName) {
-          updateMap();
-          setToggle(payload.status);
-        }
-      },
-      parentId
-    );
-
-    // remove the listener when the component unmounts
-    return () => {
-      api.event.off(EVENT_NAMES.EVENT_OVERVIEW_MAP_TOGGLE, parentId);
-    };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
-
-  const onChange = useCallback(() => {
-    updateMap();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [minimap, parentMap, zoomFactor]);
-
-  // Listen to events on the parent map
-  const handlers = useMemo(
-    () => ({ moveend: onChange, zoomend: onChange }),
-    [onChange]
-  );
-  const context = { __version: 1, map: parentMap };
-  const leafletElement = {
-    instance: parentMap,
-    context,
-  };
-  useEventHandlers(leafletElement, handlers);
-
-  return !toggle ? (
-    <div
-      style={{
-        left: `${bounds.left}px`,
-        top: `${bounds.top}px`,
-        width: `${bounds.width}px`,
-        height: `${bounds.height}px`,
-        display: "block",
-        opacity: 0.5,
-        position: "absolute",
-        border: "1px solid rgb(0, 0, 0)",
-        backgroundColor: "rgba(0, 0, 0, 0.5)",
-        zIndex: 1000,
-      }}
-    />
-  ) : null;
-}
+// get the window object
+const w = window as any;
 
 // access the cgpv object from the window object
 const cgpv = w["cgpv"];
 
 // access the api calls
-const { react, leaflet, reactLeaflet, ui } = cgpv;
+const { api, react, leaflet, reactLeaflet, ui, constants } = cgpv;
 
 const { useState, useEffect, useRef, useMemo } = react;
 
@@ -282,7 +23,11 @@ const { DomEvent } = leaflet;
 
 const { MapContainer, TileLayer, useMap } = reactLeaflet;
 
-const useStyles = ui.makeStyles((theme: any) => ({
+const { useMediaQuery, useTheme, makeStyles } = ui;
+
+const { leafletPositionClasses } = constants;
+
+const useStyles = makeStyles((theme: any) => ({
   minimap: {
     width: MINIMAP_SIZE.width,
     height: MINIMAP_SIZE.height,
@@ -306,7 +51,15 @@ const useStyles = ui.makeStyles((theme: any) => ({
   },
 }));
 
-const age = 20;
+/**
+ * Interface for overview map properties
+ */
+interface OverviewProps {
+  id: string;
+  crs: Object;
+  language: string;
+  zoomFactor: number;
+}
 
 /**
  * Create the overview map component
@@ -314,35 +67,35 @@ const age = 20;
  * @return {JSX.Element} the overview map component
  */
 export function OverviewMap(props: OverviewProps): JSX.Element {
-  const { id, crs, language, zoomFactor } = props;
+  const { id, crs, zoomFactor } = props;
 
   const [minimap, setMinimap] = useState();
 
   const classes = useStyles();
 
-  const theme = ui.useTheme();
+  const theme = useTheme();
+
+  // if screen size is medium and up
+  const deviceSizeMedUp = useMediaQuery(theme.breakpoints.up("md"));
 
   const parentMap = useMap();
   const mapZoom =
     parentMap.getZoom() - zoomFactor > 0 ? parentMap.getZoom() - zoomFactor : 0;
-  const basemaps = new Basemap(
-    { id: "transport", shaded: false, labeled: true },
-    language,
-    Number(crs.code?.split(":")[1])
-  );
+
+  const basemaps = api.map(id).basemap.getBasemapLayers();
 
   const overviewRef = useRef(null);
 
   useEffect(() => {
     // disable events on container
     const overviewHTMLElement = Cast<HTMLElement>(overviewRef.current);
-    DomEvent.disableClickPropagation(overviewHTMLElement);
-    DomEvent.disableScrollPropagation(overviewHTMLElement);
+    if (overviewHTMLElement) {
+      DomEvent.disableClickPropagation(overviewHTMLElement);
+      DomEvent.disableScrollPropagation(overviewHTMLElement);
+    }
 
     // remove ability to tab to the overview map
     // overviewHTMLElement.children[0].setAttribute("tabIndex", "-1");
-
-    console.log(age);
   }, []);
 
   // Memorize the minimap so it's not affected by position changes
@@ -372,16 +125,14 @@ export function OverviewMap(props: OverviewProps): JSX.Element {
       >
         {minimap ? (
           <>
-            {basemaps
-              .getBasemapLayers()
-              .map(
-                (base: {
-                  id: string | number | null | undefined;
-                  url: string;
-                }) => (
-                  <TileLayer key={base.id} url={base.url} />
-                )
-              )}
+            {basemaps.map(
+              (base: {
+                id: string | number | null | undefined;
+                url: string;
+              }) => (
+                <TileLayer key={base.id} url={base.url} />
+              )
+            )}
             <MinimapBounds
               parentId={id}
               parentMap={parentMap}
@@ -399,11 +150,13 @@ export function OverviewMap(props: OverviewProps): JSX.Element {
     [parentMap, crs, mapZoom, basemaps, zoomFactor, minimap]
   );
 
-  return (
-    <div className={LEAFLET_POSITION_CLASSES.topright}>
+  return deviceSizeMedUp ? (
+    <div className={leafletPositionClasses.topright}>
       <div ref={overviewRef} className="leaflet-control leaflet-bar">
-        {/* {minimapContainer} */}
+        {minimapContainer}
       </div>
     </div>
+  ) : (
+    <></>
   );
 }
